# swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "yelp-camp",
    "description": "Yelp Camp - campground project"
  },
  "host": "localhost:8080",
  "basePath": "/",
  "tags": [
    {
      "name": "YelpCamp REST API",
      "description": "Endpoints"
    }
  ],
  "schemes": [
    "http",
    "https"
  ],
  "securityDefinitions": {
    "AccessToken": {
      "type": "jwt",
      "in": "header",
      "name": "Authorization",
      "description": "Bearer accessKey"
    }
  },
  "consumes": [
    "application/x-www-form-urlencoded",
    "multipart/form-data"
  ],
  "produces": [
    "application/json",
    "text/html"
  ],
  "paths": {
    "/campgrounds": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "cover": {
                  "example": "any"
                },
                "title": {
                  "example": "any"
                },
                "price": {
                  "example": "any"
                },
                "description": {
                  "example": "any"
                },
                "location": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "302": {
            "description": "Found"
          }
        }
      }
    },
    "/campgrounds/new": {
      "get": {
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/campgrounds/api-docs": {
      "get": {
        "description": "",
        "parameters": [],
        "responses": {}
      }
    },
    "/campgrounds/{id}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "put": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "cover": {
                  "example": "any"
                },
                "title": {
                  "example": "any"
                },
                "location": {
                  "example": "any"
                },
                "price": {
                  "example": "any"
                },
                "description": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "302": {
            "description": "Found"
          },
          "400": {
            "description": "Bad Request"
          },
          "422": {
            "description": "Unprocessable Entity"
          }
        }
      },
      "delete": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "302": {
            "description": "Found"
          },
          "404": {
            "description": "Not Found"
          },
          "422": {
            "description": "Unprocessable Entity"
          }
        }
      }
    },
    "/campgrounds/{id}/edit": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "404": {
            "description": "Not Found"
          }
        }
      }
    }
  },
  "definitions": {
    "Campground": {
      "type": "object",
      "properties": {
        "title": {
          "type": "string",
          "example": "Blueberry Hills"
        },
        "price": {
          "type": "number",
          "example": 23
        },
        "cover": {
          "type": "string",
          "example": "https://source.unsplash.com/random/?city,night"
        },
        "location": {
          "type": "string",
          "example": "Kearny, New Jersey"
        },
        "slug": {
          "type": "string",
          "example": "blueberry-hills-kearny-new-jersey"
        },
        "keywords": {
          "type": "array",
          "example": [
            "Blueberry",
            "Hills",
            "Kearny",
            "New Jersey"
          ],
          "items": {
            "type": "string"
          }
        }
      }
    }
  }
}